+++ x-pack/plugins/case/server/routes/api/cases/comments/post_comment.test.ts
@@ -119,7 +119,8 @@ describe('POST comment', () => {
       createMockSavedObjectsRepository({
         caseSavedObject: mockCases,
         caseCommentSavedObject: mockCaseComments,
-      })
+      }),
+      true
     );
 
     const response = await routeHandler(theContext, request, kibanaResponseFactory);
